Update bitcoin_price.py from functions into classes
- create a folder with visualizations
- create a folder with indicators
